Moved snapshot name
